add unit test
